fake_sudir/app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Authlib==1.0.0b2
python-telegram-bot